requirements.txt

flask
pymupdf
pillow
openai
requests
